Added error handling for dictionary call
Can have a situation where there is a warc but not seed metadata, so the seed is not present in the seed_data dictionary, resulting in a KeyError and ending the script. We do not want to include a seed in the download if it has no metadata, so catching the error with try/except and continuing the script with the next seed whenever there is no metadata. Line 102-109.