Update completeness check
check_aip() for single seeds and check_aips() for batch seeds was still using the crawl start date instead of the store time date as the date for filtering for warcs that should have been included ni this download. Switched to store time to avoid problems with test crawls that aren't saved until the quarter after being crawled.