Add function make_csv(), metadata_check_combined.py
Collection and Seed reports were nearly identical code, and then added if optional or not to each which resulted in two sets of similar code within each report.

This is a start at combining all 4 versions of the reports into one function. Right now, it is the collection report updated to use parameters. Need to generalize it further.