Do not include title in the initial AIP folder name
Some web AIP titles include characters that cannot be included in a folder name, which is how the general AIP workflow gets the title into the script.

Instead, the title is saved from the Archive-It API to a dictionary, which is then used to pass the title information to preservation_xml(). Having the folder title now be the AIP ID instead of AIP ID_Title resulted in changing some other parts of the code as well.