Fix crawl job download, web_functions.py
In download_metadata(), was getting every crawl job for the entire collection. For more focused metadata, use the id to just get data on that crawl job. I don't remember why we didn't do that from the beginning. We may not have realized we could get the job id from the warc filename.