Update date variables
Renamed last_download to date_start and current_download to date_end to be more flexible now that these dates do not need to reflect the date the script is run. date_end is now a string so no longer need to convert to a string as part of reformatting but can't use datetime.date() functionality to get the year and month.