Update download_metadata() logging
To work with a dictionary of log info, have to pass log_data as an argument. To update the information, test if has the default or if need to add to existing information since data about all reports are in the same log column to keep the size of the log reasonable.