Error handling problem
Error handling was wrong - by raising an error, it stopped the function from doing the completeness check if any unexpected seeds were found and there is one in the test result. Couldn't figure out why it wasn't getting filtered out so just skip it if there is one for now.