Add seed info to metadata.csv
Need to convert department to the ARCHive group code. Otherwise, all information needed for metadata.csv is already calculated by seed_data().

Need to give MAGIL a related collection for this to work. Using 0000 so it can be handled the same as Hargrett without a related collection to not include this relationship in preservation.xml