aips_directory argument for check_aips
check_aips was assuming a specific construction for the aips_directory folder that is only true when everything from one run of the script is being made into AIPs at the same time. Sometimes, for error correction or testing, batches are split into different folders. This way is more flexible.