Create make_aip_instance
Need a function for making an AIP class specific to web AIPs so that can calculate things like title and AIP ID required for an AIP instance.

This also has some temporary changes to download_warcs() to just create a file instead of downloading because that is too slow for testing. That will be undone before the script is put into use.