Changed test order
Put the missing AIP first so that the complete AIPs could all be in a row. 2529634 is the only one with multiple warcs, so it has to be the one for that correct test. Having them in order makes it faster to review the results and see they are correct.