Use regular script function for testing
It will be more accurate if I use the regular script function. I think I can do multiple calls for the script and test the result to get input of different types. This is a start on the first test but the data types on the expected dataframe are not correct for blank values.